requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Pillow==10.0.1
requests==2.31.0
aiohttp==3.8.6
openai==1.3.7
python-socketio==5.9.0
python-engineio==4.7.1
